examples/pydantic-ai/pyproject.toml
[project]
name = "pydantic-ai-example"
version = "0.1.0"
description = "Build a typed browser agent with Pydantic AI and Steel"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai>=1.0.0",
    "playwright>=1.48.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "steel-sdk>=0.16.0",
]
